danesfield/segmentation/semantic/requirements.txt
PyTorch
pencv-python
sklearn
scikit-image
ubelt
tensorboardX
sympy
tqdm
numpy
scipy
gdal
pdal
